Renaming serializer.
