+++ shared/src/commonMain/kotlin/com/adammcneilly/pwhl/mobile/shared/models/PlayByPlayEvent.kt
@@ -1,0 +1,5 @@
+package com.adammcneilly.pwhl.mobile.shared.models
+
+data class PlayByPlayEvent(
+    val eventType: String,
+)
